+++ backend/src/plugins/Automod/commands/AntiraidClearCmd.ts
@@ -5,6 +5,7 @@ import { AutomodPluginType } from "../types";
 
 export const AntiraidClearCmd = guildPluginMessageCommand<AutomodPluginType>()({
   trigger: ["antiraid clear", "antiraid reset", "antiraid none", "antiraid off"],
+  description: "Turns antiraid off",
   permission: "can_set_antiraid",
 
   async run({ pluginData, message }) {