+++ backend/src/plugins/Automod/commands/ViewAntiraidCmd.ts
@@ -4,6 +4,7 @@ import { AutomodPluginType } from "../types";
 export const ViewAntiraidCmd = guildPluginMessageCommand<AutomodPluginType>()({
   trigger: "antiraid",
   permission: "can_view_antiraid",
+  description: "Sets the antiraid level",
 
   async run({ pluginData, message }) {
     if (pluginData.state.cachedAntiraidLevel) {