+++ backend/src/plugins/BotControl/commands/AddDashboardUserCmd.ts
@@ -18,7 +18,7 @@ export const AddDashboardUserCmd = botControlCmd({
   async run({ pluginData, message: msg, args }) {
     const guild = await pluginData.state.allowedGuilds.find(args.guildId);
     if (!guild) {
-      sendErrorMessage(pluginData, msg.channel, "Server is not using Zeppelin");
+      sendErrorMessage(pluginData, msg.channel, "Server is not using Streetcar");
       return;
     }