+++ backend/src/plugins/BotControl/commands/RemoveDashboardUserCmd.ts
@@ -17,7 +17,7 @@ export const RemoveDashboardUserCmd = botControlCmd({
   async run({ pluginData, message: msg, args }) {
     const guild = await pluginData.state.allowedGuilds.find(args.guildId);
     if (!guild) {
-      sendErrorMessage(pluginData, msg.channel, "Server is not using Zeppelin");
+      sendErrorMessage(pluginData, msg.channel, "Server is not using Streetcar");
       return;
     }